Update path references to Ubigraph since it was relocated in pull request #248 and BuildBot is failing
